+++ R/utilities.R
@@ -194,6 +194,7 @@ CellsPerGroup <- function(
 #' @importFrom S4Vectors subjectHits mcols
 #' @importFrom methods is
 #' @importFrom Seurat DefaultAssay
+#' @importFrom GenomeInfoDb dropSeqlevels
 #' @return Returns a dataframe with the name of each region, the closest feature
 #' in the annotation, and the distance to the feature.
 #' @export
@@ -222,7 +223,28 @@ ClosestFeature <- function(
   if (!inherits(x = object, what = "ChromatinAssay")) {
     stop("The requested assay is not a ChromatinAssay.")
   }
+  if (length(x = regions) == 0) {
+    stop("No query regions supplied")
+  }
   annotation <- SetIfNull(x = annotation, y = Annotation(object = object))
+  missing_seqlevels <- setdiff(
+    x = seqlevels(x = regions), y = seqlevels(x = annotation)
+  )
+  if (length(x = missing_seqlevels) > 0) {
+    warning(
+      "The following seqlevels present in query regions are not present\n ",
+      "in the supplied gene annotations and will be removed: ",
+      paste(missing_seqlevels, collapse = ", ")
+    )
+    regions <- dropSeqlevels(
+      x = regions,
+      value = missing_seqlevels,
+      pruning.mode = "coarse"
+    )
+    if (length(x = regions) == 0) {
+      stop("None of the supplied regions were found in the supplied annotation")
+    }
+  }
   nearest_feature <- distanceToNearest(x = regions, subject = annotation)
   feature_hits <- annotation[subjectHits(x = nearest_feature)]
   df <- as.data.frame(x = mcols(x = feature_hits))
@@ -276,6 +298,11 @@ GeneActivity <- function(
   verbose = TRUE,
   ...
 ) {
+  if (!is.null(x = features)) {
+    if (length(x = features) == 0) {
+      stop("Empty list of features provided")
+    }
+  }
   # collapse to longest protein coding transcript
   assay <- SetIfNull(x = assay, y = DefaultAssay(object = object))
   if (!inherits(x = object[[assay]], what = "ChromatinAssay")) {
@@ -291,15 +318,24 @@ GeneActivity <- function(
   transcripts <- CollapseToLongestTranscript(ranges = annotation)
   if (!is.null(x = biotypes)) {
     transcripts <- transcripts[transcripts$gene_biotype %in% biotypes]
+    if (length(x = transcripts) == 0) {
+      stop("No genes remaining after filtering for requested biotypes")
+    }
   }
 
   # filter genes if provided
   if (!is.null(x = features)) {
     transcripts <- transcripts[transcripts$gene_name %in% features]
+    if (length(x = transcripts) == 0) {
+      stop("None of the requested genes were found in the gene annotation")
+    }
   }
   if (!is.null(x = max.width)) {
     transcript.keep <- which(x = width(x = transcripts) < max.width)
     transcripts <- transcripts[transcript.keep]
+    if (length(x = transcripts) == 0) {
+      stop("No genes remaining after filtering for max.width")
+    }
   }
 
   # extend to include promoters
@@ -780,7 +816,9 @@ LookupGeneCoords <- function(object, gene, assay = NULL) {
     stop("The requested assay is not a ChromatinAssay")
   }
   annotations <- Annotation(object = object[[assay]])
-  annot.sub <- annotations[annotations$gene_name == gene]
+  isgene <- annotations$gene_name == gene
+  isgene <- !is.na(x = isgene) & isgene
+  annot.sub <- annotations[isgene]
   if (length(x = annot.sub) == 0) {
     return(NULL)
   } else {
@@ -1206,15 +1244,22 @@ FindRegion <- function(
   extend.downstream = 0
 ) {
   if (!is(object = region, class2 = "GRanges")) {
-    # if separators are present in the string and we can convert the
-    # start to a number, assume we're using genomic coordinates
-    if (all(sapply(X = sep, FUN = grepl, x = region))) {
-      region <- StringToGRanges(regions = region, sep = sep)
-    } else {
-      region <- LookupGeneCoords(object = object, assay = assay, gene = region)
-      if (is.null(x = region)) {
-        stop("Gene not found")
+    # first try to convert to coordinates, if not lookup gene
+    region <- tryCatch(
+      expr = suppressWarnings(
+        expr = StringToGRanges(regions = region, sep = sep)
+      ),
+      error = function(x) {
+        region <- LookupGeneCoords(
+          object = object,
+          assay = assay,
+          gene = region
+        )
+        return(region)
       }
+    )
+    if (is.null(x = region)) {
+      stop("Gene not found")
     }
   }
   region <- suppressWarnings(expr = Extend(
@@ -1274,7 +1319,6 @@ GetReadsInRegion <- function(
     pruning.mode = "coarse"
   )
   reads <- scanTabix(file = tabix.file, param = region)
-  invisible(x = gc(verbose = FALSE))
   reads <- TabixOutputToDataFrame(reads = reads)
   reads <- reads[
     fmatch(x = reads$cell, table = cellmap, nomatch = 0L) > 0,
@@ -1742,6 +1786,17 @@ TabixOutputToDataFrame <- function(reads, record.ident = TRUE) {
     nrep <- elementNROWS(x = reads)
   }
   reads <- unlist(x = reads, use.names = FALSE)
+  if (length(x = reads) == 0) {
+    df <- data.frame(
+      "chr" = "",
+      "start" = "",
+      "end" = "",
+      "cell" = "",
+      "count" = ""
+    )
+    df <- df[-1, ]
+    return(df)
+  }
   reads <- stri_split_fixed(str = reads, pattern = "\t")
   n <- length(x = reads[[1]])
   unlisted <- unlist(x = reads)