+++ NEWS.md
@@ -1,3 +1,23 @@
+# Signac 1.4.0
+
+Bug fixes:
+
+* Fixed bug in `FindMotifs()` when using only one region as input ([#732](https://github.com/timoast/signac/issues/732))
+* Add check for correct number of columns in fragment file ([#748](https://github.com/timoast/signac/issues/748))
+* Fixed gene lookup when annotations contain NA values ([#771](https://github.com/timoast/signac/issues/771))
+* Fixed error in `ClosestFeature()` when query contained regions on contigs not present in gene annotation ([#758](https://github.com/timoast/signac/issues/758))
+* Fixed bug in `TSSEnrichment()` when using multiple fragment files ([#783](https://github.com/timoast/signac/issues/783))
+* Fixed bug in `CallPeaks()` when multiple fragment files used as input
+* Fixed bug in `CallPeaks()` to account for 0-based starts in called peaks
+* Fixed bug in gene name lookup when gene names contain `-` characters ([#759](https://github.com/timoast/signac/issues/759))
+
+Other changes:
+
+* Updated documentation for `genome` parameter in `AddMotifs()` and `RunChromVAR()` ([#712](https://github.com/timoast/signac/issues/712))
+* Updated the `FoldChange()` function to use normalized counts rather than raw counts ([#795](https://github.com/timoast/signac/issues/795))
+* Improved error checking in `GeneActivity()` ([#797](https://github.com/timoast/signac/issues/797))
+* Added `format` parameter to `CallPeaks()` ([#682](https://github.com/timoast/signac/issues/682))
+
 # Signac 1.3.0 
 
 Bug fixes: